+++ src/routes/authRouter.js
@@ -1,6 +1,6 @@
 import express from "express";
-import { authValidation } from "../middlewares/authorization.middleware.js";
-import * as authController from "../controllers/authorization.controllers.js";
+import { authValidation } from "../middlewares/authMiddleware.js";
+import * as authController from "../controllers/authController.js";
 
 const router = express.Router();
